requirements.txt
rpi.gpio
spidev
gpiozero
Flask~=3.0.3
filelock~=3.15.4
pillow~=10.4.0
